+++ app/src/main/java/com/cajetan/youtubeplus/utils/YouTubeData.kt
@@ -126,8 +126,7 @@ class YouTubeData(parentActivity: Activity, fragment: Fragment? = null) :
 
     private fun videoSearchTask(videoId: String, pageToken: String) {
         doAsync {
-            val result: List<Video>
-            val result2: MutableList<FeedItem> = emptyList()
+            val result: ArrayList<FeedItem> = ArrayList()
             val nextPageToken: String
             val prevPageToken: String
 
@@ -176,21 +175,21 @@ class YouTubeData(parentActivity: Activity, fragment: Fragment? = null) :
                 if (finalChannelIds.isNotEmpty())
                     finalChannelIds.setLength(finalChannelIds.length - 1)
 
-                result = service.videos().list("snippet,contentDetails")
+                result.addAll(service.videos().list("snippet,contentDetails")
                         .setId(finalVideoIds.toString()).execute().items
+                        .map { FeedItem(it.id, video = it) })
 
-                result2.addAll(service.channels().list("snippet,contentDetails")
-                        .setId(finalChannelIds.toString()).execute().items
-                        .map { FeedItem(it.id, channel = it) })
+//                result.addAll(service.channels().list("snippet,contentDetails")
+//                        .setId(finalChannelIds.toString()).execute().items
+//                        .map { FeedItem(it.id, channel = it) })
 
-                result2.addAll(service.playlists().list("snippet,contentDetails")
-                        .setId(finalPlaylistIds.toString()).execute().items
-                        .map { FeedItem(it.id, playlist = it) })
-
-                result2.addAll(result.map { FeedItem(it.id, video = it) })
+//                result.addAll(service.playlists().list("snippet,contentDetails")
+//                        .setId(finalPlaylistIds.toString()).execute().items
+//                        .map { FeedItem(it.id, playlist = it) })
 
             } catch (e: java.lang.Exception) {
                 onTaskCancelled(e)
+                Log.d("YouTubeData", "Exception $e")
                 throw CancellationException()
             }
 
@@ -459,7 +458,7 @@ class YouTubeData(parentActivity: Activity, fragment: Fragment? = null) :
     }
 
     interface VideoSearchListener {
-        fun onSearchResultsReceived(results: List<Video>,
+        fun onSearchResultsReceived(results: List<FeedItem>,
                                     nextPageToken: String, previousPageToken: String)
     }